refactor(batch-exports): split temporal/sql.py into per-model modules
`temporal/sql.py` was 991 lines covering three batch export models (events,
persons, sessions) across two query families: `SELECT_FROM_*`, which reads via
ArrowStream, and `EXPORT_TO_S3_*`, which wraps the same query in an
`INSERT INTO FUNCTION s3(...)`. Hard to navigate and hard to review.

Split it into a `temporal/sql/` package, one module per model:

    sql/common.py     get_s3_function_call, HogQLQueryBatchExportSettings
    sql/events.py     6 SELECT + 6 EXPORT pairs, EVENT_COUNT_BY_INTERVAL
    sql/persons.py    2 SELECT + 2 EXPORT pairs
    sql/sessions.py   SELECT_FROM_SESSIONS_HOGQL

Grouped by model rather than by family, so each `SELECT_FROM_X` sits directly
next to the `EXPORT_TO_S3_FROM_X` that queries the same source table. The two
are near-identical apart from the INSERT wrapper, and three of the eight pairs
have already quietly diverged; keeping them adjacent makes a divergence visible
on sight. Splitting by family would have done the opposite.

`HogQLQueryBatchExportSettings` goes in `common`, not `sessions`: it sat next to
the sessions query in the old file, but `record_batch_model` also uses it for the
HogQL query model.

No SQL was changed. Every template body was extracted by line range and verified
byte-identical to the original, including the `SELECT_FROM_SESSIONS_HOGQL` AST
literal. `__init__.py` deliberately re-exports nothing, so each call site
declares which model's SQL it depends on.

Generated-By: PostHog Code
Task-Id: e2c0e744-189b-4ac0-a1f4-c999b0a0e071
