+++ products/batch_exports/backend/temporal/monitoring.py
@@ -14,7 +14,7 @@ from posthog.temporal.common.logger import get_logger
 
 from products.batch_exports.backend.models.batch_export import BatchExport, BatchExportRun
 from products.batch_exports.backend.service import afetch_batch_export_runs_in_range, aupdate_records_total_count
-from products.batch_exports.backend.temporal.sql import EVENT_COUNT_BY_INTERVAL
+from products.batch_exports.backend.temporal.sql.events import EVENT_COUNT_BY_INTERVAL
 
 LOGGER = get_logger(__name__)
